backend/requirements.txt
Django==3.2.16
Pillow==9.3.0
djangorestframework==3.14.0
djoser==2.1.0
django-cors-headers==4.2.0
drf-extra-fields==3.5.0
django-filter==22.1
python-dotenv==1.0.0
psycopg2-binary==2.9.3
flake8
 